Add Typescript + Add Tests
