+++ public/index.html
@@ -13,7 +13,7 @@
       content="Web site created using create-react-app"
     />
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+    <!--<link rel="manifest" href="%PUBLIC_URL%/manifest.json" />-->
     <title>Wordle</title>
   </head>
   <body>
