+++ public/manifest.json
@@ -1,6 +1,5 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "name": "Wordle",
   "icons": [
     {
       "src": "favicon.ico",
