--- src/axios.js
@@ -1,8 +1,0 @@
-import axios from 'axios';
-
-const instance = axios.create({
-  baseURL: 'https://api.dictionaryapi.dev/api/v2/entries/en/',
-  method: 'GET',
-});
-
-export default instance;
