+++ src/components/KeyBoardKey/KeyBoardKey.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import './KeyBoardKey.css'
-const KeyBoardKey = ({kbkey, className, onClick }) => {
+const KeyBoardKey = ({ kbkey, className, onClick } : { kbkey : string, className : string, onClick : React.MouseEventHandler<HTMLDivElement> | undefined }) => {
   return <div className={`key ${className}`} onClick={onClick}>
     { kbkey }
     </div>;
