--- src/components/modal/modal.test.js
@@ -1,10 +1,0 @@
-import React from "react";
-import { shallow } from "enzyme";
-import Modal from "./modal";
-
-describe("Modal", () => {
-  test("matches snapshot", () => {
-    const wrapper = shallow(<Modal />);
-    expect(wrapper).toMatchSnapshot();
-  });
-});
