+++ src/components/modal/modal.tsx
@@ -1,5 +1,7 @@
+import React from 'react';
 import './modal.css';
-const Modal = ({message}) => {
+
+const Modal = ({message} : { message : string}) => {
   return <div className="modal">
     {message}
     </div>;
