+++ src/components/navbar/navbar.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import './navbar.css'
 const Navbar = () => {
   return <div className="navbar">
-              <span className="logo">Wordle</span>
-          </div>;
+    <span className="logo">Wordle</span>
+  </div>;
 };
 
 export default Navbar;
